feat: added the cost optimization of instancetype across region
